backend/requirements.txt
# Flask项目依赖包
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
pymysql==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0